Add base object types to repository
This commit adds base object types to the repository definitions. These
include base_applications, base_modifiers, and base_package_managers.
The purpose of the base objects is to separate non-working but
inheritable base definitions of objects from their working counterparts.
